Improve install after removing deployment setup
The Metalware 'deployment' directory has now been removed, and its
contents refactored to this repo (for code related to building HPC
nodes) or Imageware (for code related to setting up a Metalware
deployment server).

This means this installation script can/must be improved in two ways:

- Simplify the installation functions to only handle the case of
installation from a remote server, as this repo does not need to handle
setting up the current node as a Metalware deployment server (this now
happens in Imageware).

- Remove the assumption that Metalware will have the 'deployment'
directory, and the deployment server will serve this. This only happened
to work for now as the deployment directory had the same needed files as
this repo, however during the installation this repo's scripts should
actually request files from the same repo.