Replace `buildserver` parameter with `hostip`
Now that `hostip` should be more reliably detected as the IP of the
deployment server on the `prv`/build network there is no need to have a
separate parameter that must be set to define this, we can just use
`hostip` and this should Just Work.

See also: https://github.com/alces-software/metalware/pull/45.